feat: move fixtures
