+++ tests/test_results.py
@@ -1,32 +1,9 @@
 import pandas as pd
 import numpy as np
 import pytest
-from numpy.ma.extras import column_stack
 
 from pyneapple import Results
 from pyneapple import Parameters, IVIMParams
-
-
-@pytest.fixture
-def random_results(ivim_tri_params):
-    f = {(0, 0, 0): [1.1, 1.2, 1.3]}
-    d = {(0, 0, 0): [1.0, 1.2, 1.3]}
-    s_0 = {(0, 0, 0): np.random.rand(1)}
-    results = Results(ivim_tri_params)
-    results.f.update(f)
-    results.d.update(d)
-    results.s_0.update(s_0)
-    return results
-
-
-@pytest.fixture
-def array_result():
-    """Random decay signal."""
-    spectrum = np.zeros((2, 2, 1, 11))
-    bins = np.linspace(0, 10, 11)
-    for index in np.ndindex((2, 2)):
-        spectrum[index] = np.exp(-np.kron(bins, abs(np.random.randn(1))))
-    return spectrum
 
 
 def test_custom_dict_validate_key():
